backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
apscheduler>=3.10.0
feedparser>=6.0.10
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9